pyproject.toml
[project]
name = "project-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Huy Duong",email = "duongquanghuy2202@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "sentence-transformers (>=5.2.3,<6.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "langchain-classic (>=1.0.1,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
